fix: Reset password by email failed (#595)
* fix: ResetPasswordReq pb

* fix: reset psw error

* Revert "fix: ResetPasswordReq pb"

This reverts commit 7e884dc225a523e219333332859ec236a2b3fdd1.

+++ pkg/common/db/database/chat.go
@@ -255,6 +255,9 @@ func (o *ChatDatabase) UpdatePasswordAndDeleteVerifyCode(ctx context.Context, us
 		if err := o.account.UpdatePassword(ctx, userID, password); err != nil {
 			return err
 		}
+		if codeID == "" {
+			return nil
+		}
 		if err := o.verifyCode.Delete(ctx, codeID); err != nil {
 			return err
 		}
